Extracted additional traits
