+++ core/Request.php
@@ -1,14 +1,18 @@
 <?php namespace spitfire\core;
 
 use magic3w\http\url\reflection\URLReflection;
-use spitfire\io\Get;
 use spitfire\exceptions\ApplicationException;
 use Psr\Http\Message\ServerRequestInterface;
 use Psr\Http\Message\StreamInterface;
 use Psr\Http\Message\UriInterface;
 use Psr\Http\Message\UploadedFileInterface;
+use spitfire\core\http\request\components\hasBody;
 use spitfire\core\http\request\components\hasHTTPMethod;
+use spitfire\core\http\request\components\hasParsedBody;
 use spitfire\core\http\request\components\hasProtocolVersion;
+use spitfire\core\http\request\components\hasServerParams;
+use spitfire\core\http\request\components\hasUploads;
+use spitfire\core\http\request\components\hasURI;
 use spitfire\io\stream\Stream;
 use spitfire\io\UploadFile;
 
@@ -38,58 +42,14 @@ use spitfire\io\UploadFile;
  */
 class Request implements ServerRequestInterface
 {
-	use hasProtocolVersion, hasHTTPMethod;
-	
-	/**
-	 * Contains data that can be retrieved out of the Path string of the URL. For
-	 * example, the controller / action / object of the request or data about
-	 * custom route parameters.
-	 *
-	 * @var UriInterface
-	 */
-	private $uri;
-	
-	/**
-	 * The server parameter array contains all the runtime information the server
-	 * has about itself and the current request. Usually this data is directly
-	 * originated from $_SERVER
-	 *
-	 * @var string[]
-	 */
-	private $serverParams;
-	
-	/**
-	 * Contains information what information was delivered via query parameters of
-	 * the URL. This will be a special _GET object that will aid generating canonicals.
-	 * If you're writing an application/component that cannot guarantee having the
-	 * <i>request.replace_globals</i> enabled you should use this rather than $_GET
-	 *
-	 * @var \spitfire\io\Get
-	 */
-	private $get;
-	
-	/**
-	 * The array tree of uploads, this is a mix of UploadInterfaces and arrays. The
-	 * application can the use the array to manipulate the uploaded file.
-	 *
-	 * @see https://www.php-fig.org/psr/psr-7/meta/
-	 * @var (array|UploadedFileInterface)[]
-	 */
-	private $uploads;
-	
-	/**
-	 * This contains a mixed version of _POST and _FILES that allows your app to
-	 * conveniently use the data generated by this two sources in a single place.
-	 *
-	 * @var mixed
-	 */
-	private $post;
-	
-	/**
-	 *
-	 * @var StreamInterface
-	 */
-	private $body;
+	use
+	hasProtocolVersion,
+	hasHTTPMethod,
+	hasURI,
+	hasServerParams,
+	hasUploads,
+	hasParsedBody,
+	hasBody;
 	
 	/**
 	 *
@@ -152,91 +112,6 @@ class Request implements ServerRequestInterface
 	}
 	
 	/**
-	 * The server params function returns the equivalent of the _SERVER
-	 * superglobal. We generally set the content of the variable to _SERVER
-	 * but you may come across implementations that override it completely
-	 * or partially.
-	 *
-	 * @return string[]
-	 */
-	public function getServerParams()
-	{
-		return $this->serverParams;
-	}
-	
-	/**
-	 * Returns a tree of uploaded files. The tree contains arrays mixed with UploadedFileInterface
-	 * objects that contain the actual metadata.
-	 *
-	 * @see https://www.php-fig.org/psr/psr-7/meta/
-	 * @return (array|UploadedFileInterface)[]
-	 */
-	public function getUploadedFiles()
-	{
-		return $this->uploads;
-	}
-	
-	/**
-	 * Returns the request target. I am still not 100% certain what the request target is, it seems
-	 * to be some kind of fallback that I am not certain how it works (apparently you can request
-	 * wildcard data from some servers).
-	 *
-	 * @return string
-	 */
-	public function getRequestTarget() : string
-	{
-		if ($this->uri === null) {
-			return '/';
-		}
-		
-		$query = $this->uri->getQuery();
-		return $this->uri->getPath() . ($query? '?' . $query: '');
-	}
-	
-	/**
-	 * Returns the URI used to request a resource from the server.
-	 *
-	 * @return UriInterface
-	 */
-	public function getUri() : UriInterface
-	{
-		return $this->uri;
-	}
-	
-	/**
-	 * In our service this is not implemented. Calling it has no effect
-	 */
-	public function withRequestTarget($requestTarget) : Request
-	{
-		return $this;
-	}
-	
-	/**
-	 * Returns the array of query parameters (_GET) that the request was sent along
-	 * with.
-	 *
-	 * @return mixed[]
-	 */
-	public function getQueryParams()
-	{
-		if ($this->get === null) {
-			parse_str($this->uri->getQuery(), $get);
-			$this->get = new Get($get);
-		}
-		
-		return $this->get->getRaw();
-	}
-	
-	/**
-	 *
-	 * @return mixed[]
-	 */
-	public function getParsedBody()
-	{
-		return $this->post;
-	}
-	
-	/**
 	 *
 	 * @return mixed[]
 	 */
@@ -259,63 +134,9 @@ class Request implements ServerRequestInterface
 		return $this->cookie;
 	}
 	
-	/**
-	 * Returns the content that is to be sent with the body. This is a string your
-	 * application has to set beforehand.
-	 *
-	 * In case you're using Spitfire's Context object to manage the context the
-	 * response will get the view it contains and render that before returning it.
-	 *
-	 * @return StreamInterface
-	 */
-	public function getBody() : StreamInterface
-	{
-		return $this->body;
-	}
-	
-	public function withBody(StreamInterface $body)
-	{
-		$copy = clone $this;
-		$copy->body = $body;
-		return $copy;
-	}
-	
 	public function getHeaders()
 	{
 		return $this->headers->all();
-	}
-	
-	/**
-	 * Override the request URI for the server.
-	 *
-	 * @return ServerRequestInterface
-	 */
-	public function withUri(UriInterface $uri, $preserveHost = false)
-	{
-		if ($preserveHost) {
-			if (!$this->hasHeader('Host') && $uri->getHost()) {
-				$copy = $this->withHeader('Host', $uri->getHost());
-			}
-			else {
-				$copy = clone $this;
-			}
-		}
-		else {
-			$copy = $this->withHeader('Host', $uri->getHost());
-		}
-		
-		
-		$copy->uri = $uri;
-		
-		return $copy;
-	}
-	
-	public function withMethod($method)
-	{
-		$copy = clone $this;
-		$copy->method = $method;
-		
-		return $copy;
 	}
 	
 	public function withProtocolVersion($version)
@@ -323,30 +144,6 @@ class Request implements ServerRequestInterface
 		$copy = clone $this;
 		$copy->version = $version;
 		
-		return $copy;
-	}
-	
-	/**
-	 *
-	 * @param mixed[] $query
-	 * @return Request
-	 */
-	public function withQueryParams(array $query) : Request
-	{
-		$copy = clone $this;
-		$copy->get = new Get($query);
-		
-		return $copy;
-	}
-	
-	/**
-	 *
-	 * @param null|mixed[] $data
-	 */
-	public function withParsedBody($data) : Request
-	{
-		$copy = clone $this;
-		$copy->post = $data;
 		return $copy;
 	}
 	
@@ -480,27 +277,6 @@ class Request implements ServerRequestInterface
 	}
 	
 	/**
-	 *
-	 * @param (array|UploadedFileInterface)[] $uploadedFiles
-	 */
-	public function withUploadedFiles(array $uploadedFiles)
-	{
-		$copy = clone $this;
-		$copy->uploads = $uploadedFiles;
-		return $copy;
-	}
-	
-	/**
-	 * Whether this request was posted.
-	 *
-	 * @return bool
-	 */
-	public function isPost() : bool
-	{
-		return $this->method === 'POST';
-	}
-	
-	/**
 	 * If the request contains range information (the client only wishes to retrieve a subset of the
 	 * resource), this endpoint will return true.
 	 *
@@ -564,7 +340,7 @@ class Request implements ServerRequestInterface
 		 */
 		$uri = URLReflection::fromGlobals();
 		
-		$uploads = UploadFile::fromGlobals();
+		$uploads = self::filesFromGlobals();
 		
 		/**
 		 * The PSR strongly suggests that both the cookie and server variable should be directly
