+++ io/UploadFile.php
@@ -205,38 +205,4 @@ class UploadFile implements UploadedFileInterface
 		
 		return $sizes[0];
 	}
-	
-	/**
-	 *
-	 * @todo Support nested file arrays.
-	 */
-	public static function fromGlobals()
-	{
-		$_return = [];
-		
-		foreach ($_FILES as $key => $upload) {
-			
-			/**
-			 * We currently do not support uploads that are within nested arrays. This is a
-			 * known limitation, but it should be fine for our use case.
-			 *
-			 * We do not assert this, since it is entirely possible that a user submits bad
-			 * data during runtime.
-			 */
-			assume(!is_array($upload['name']), 'File array support is currently disabled');
-			
-			/**
-			 * Add the uploaded file to the list of uploads.
-			 */
-			$_return[$key] = new UploadFile(
-				$upload['tmp_name'],
-				$upload['name'],
-				$upload['type'],
-				$upload['size'],
-				$upload['error']
-			);
-		}
-		
-		return $_return;
-	}
 }
